+++ jib-cli/src/main/java/com/google/cloud/tools/jib/cli/JibCli.java
@@ -19,15 +19,21 @@ package com.google.cloud.tools.jib.cli;
 import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.apache.v2.ApacheHttpTransport;
 import com.google.cloud.tools.jib.ProjectInfo;
+import com.google.cloud.tools.jib.api.JibContainer;
 import com.google.cloud.tools.jib.api.LogEvent;
 import com.google.cloud.tools.jib.cli.logging.Verbosity;
+import com.google.cloud.tools.jib.plugins.common.ImageMetadataOutput;
 import com.google.cloud.tools.jib.plugins.common.UpdateChecker;
 import com.google.cloud.tools.jib.plugins.common.globalconfig.GlobalConfig;
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLogger;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.Futures;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.Optional;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -115,6 +121,24 @@ public class JibCli {
                       privacyUrl);
               logger.log(LogEvent.Level.LIFECYCLE, message);
             });
+  }
+
+  /**
+   * Writes image details (imageId, digest, tags, etc.) to a json file, if the path to the json is
+   * provided.
+   *
+   * @param imageJsonOutputPath optional path to json file (for example,
+   *     path/to/json/jib-image.json)
+   * @param jibContainer the {@link JibContainer} to derive image details from
+   * @throws IOException if error occurs when writing to the json file.
+   */
+  static void writeImageJson(Optional<Path> imageJsonOutputPath, JibContainer jibContainer)
+      throws IOException {
+    if (imageJsonOutputPath.isPresent()) {
+      ImageMetadataOutput metadataOutput = ImageMetadataOutput.fromJibContainer(jibContainer);
+      Files.write(
+          imageJsonOutputPath.get(), metadataOutput.toJson().getBytes(StandardCharsets.UTF_8));
+    }
   }
 
   /**