+++ jib-cli/src/test/java/com/google/cloud/tools/jib/cli/JibCliTest.java
@@ -25,12 +25,23 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 import com.google.cloud.tools.jib.ProjectInfo;
+import com.google.cloud.tools.jib.api.DescriptorDigest;
+import com.google.cloud.tools.jib.api.ImageReference;
+import com.google.cloud.tools.jib.api.InvalidImageReferenceException;
+import com.google.cloud.tools.jib.api.JibContainer;
 import com.google.cloud.tools.jib.api.LogEvent;
 import com.google.cloud.tools.jib.cli.logging.Verbosity;
+import com.google.cloud.tools.jib.json.JsonTemplateMapper;
+import com.google.cloud.tools.jib.plugins.common.ImageMetadataOutput;
 import com.google.cloud.tools.jib.plugins.common.globalconfig.GlobalConfig;
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLogger;
+import com.google.common.collect.ImmutableSet;
 import com.google.common.util.concurrent.Futures;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.security.DigestException;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
@@ -38,7 +49,9 @@ import java.util.logging.ConsoleHandler;
 import java.util.logging.Handler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
@@ -48,6 +61,10 @@ public class JibCliTest {
 
   @Mock private GlobalConfig globalConfig;
   @Mock private ConsoleLogger logger;
+
+  @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+  @Mock private JibContainer mockJibContainer;
 
   @Test
   public void testConfigureHttpLogging() {
@@ -103,5 +120,27 @@ public class JibCliTest {
                     + "). Download the latest Jib CLI version from "
                     + ProjectInfo.GITHUB_URL
                     + "/releases/tag/v2.0.0-cli"));
+  }
+
+  @Test
+  public void testWriteImageJson()
+      throws InvalidImageReferenceException, IOException, DigestException {
+    String imageId = "sha256:61bb3ec31a47cb730eb58a38bbfa813761a51dca69d10e39c24c3d00a7b2c7a9";
+    String digest = "sha256:3f1be7e19129edb202c071a659a4db35280ab2bb1a16f223bfd5d1948657b6fc";
+    when(mockJibContainer.getTargetImage()).thenReturn(ImageReference.parse("adoptopenjdk:8-jre"));
+    when(mockJibContainer.getImageId()).thenReturn(DescriptorDigest.fromDigest(imageId));
+    when(mockJibContainer.getDigest()).thenReturn(DescriptorDigest.fromDigest(digest));
+    when(mockJibContainer.getTags()).thenReturn(ImmutableSet.of("latest", "tag-2"));
+
+    Path outputPath = temporaryFolder.getRoot().toPath().resolve("jib-image.json");
+    JibCli.writeImageJson(Optional.of(outputPath), mockJibContainer);
+
+    String outputJson = new String(Files.readAllBytes(outputPath), StandardCharsets.UTF_8);
+    ImageMetadataOutput metadataOutput =
+        JsonTemplateMapper.readJson(outputJson, ImageMetadataOutput.class);
+    assertThat(metadataOutput.getImage()).isEqualTo("adoptopenjdk:8-jre");
+    assertThat(metadataOutput.getImageId()).isEqualTo(imageId);
+    assertThat(metadataOutput.getImageDigest()).isEqualTo(digest);
+    assertThat(metadataOutput.getTags()).containsExactly("latest", "tag-2");
   }
 }