+++ jib-plugins-common/src/test/java/com/google/cloud/tools/jib/plugins/common/UpdateCheckerTest.java
@@ -16,7 +16,10 @@
 
 package com.google.cloud.tools.jib.plugins.common;
 
-import com.google.cloud.tools.jib.api.LogEvent.Level;
+import static com.google.common.truth.Truth.assertThat;
+import static com.google.common.truth.Truth8.assertThat;
+
+import com.google.cloud.tools.jib.api.LogEvent;
 import com.google.cloud.tools.jib.http.TestWebServer;
 import com.google.common.util.concurrent.Futures;
 import java.io.IOException;
@@ -31,10 +34,7 @@ import java.time.Instant;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.Future;
-import org.hamcrest.CoreMatchers;
-import org.hamcrest.MatcherAssert;
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -74,13 +74,13 @@ public class UpdateCheckerTest {
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "1.0.2", testWebServer.getEndpoint(), "tool-name", ignored -> {});
-    Assert.assertTrue(testWebServer.getInputRead().contains("User-Agent: jib 1.0.2 tool-name"));
-    Assert.assertTrue(message.isPresent());
-    Assert.assertEquals("2.0.0", message.get());
+
+    assertThat(testWebServer.getInputRead()).contains("User-Agent: jib 1.0.2 tool-name");
+    assertThat(message).hasValue("2.0.0");
     String modifiedTime =
         new String(
             Files.readAllBytes(configDir.resolve("lastUpdateCheck")), StandardCharsets.UTF_8);
-    Assert.assertTrue(Instant.parse(modifiedTime).isAfter(before));
+    assertThat(Instant.parse(modifiedTime)).isGreaterThan(before);
   }
 
   @Test
@@ -93,8 +93,8 @@ public class UpdateCheckerTest {
       Optional<String> message =
           UpdateChecker.performUpdateCheck(
               configDir, "1.0.2", server.getEndpoint(), "tool-name", ignored -> {});
-      Assert.assertTrue(message.isPresent());
-      Assert.assertEquals("2.0.0", message.get());
+
+      assertThat(message).hasValue("2.0.0");
     }
   }
 
@@ -105,12 +105,13 @@ public class UpdateCheckerTest {
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "2.0.0", testWebServer.getEndpoint(), "tool-name", ignored -> {});
-    Assert.assertFalse(message.isPresent());
+
+    assertThat(message).isEmpty();
     String modifiedTime =
         new String(
             Files.readAllBytes(configDir.resolve("lastUpdateCheck")), StandardCharsets.UTF_8);
-    Assert.assertTrue(testWebServer.getInputRead().contains("User-Agent: jib 2.0.0 tool-name"));
-    Assert.assertTrue(Instant.parse(modifiedTime).isAfter(before));
+    assertThat(testWebServer.getInputRead()).contains("User-Agent: jib 2.0.0 tool-name");
+    assertThat(Instant.parse(modifiedTime)).isGreaterThan(before);
   }
 
   @Test
@@ -119,12 +120,12 @@ public class UpdateCheckerTest {
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "1.0.2", testWebServer.getEndpoint(), "tool-name", ignored -> {});
-    Assert.assertTrue(message.isPresent());
-    Assert.assertEquals("2.0.0", message.get());
+
+    assertThat(message).hasValue("2.0.0");
     String modifiedTime =
         new String(
             Files.readAllBytes(configDir.resolve("lastUpdateCheck")), StandardCharsets.UTF_8);
-    Assert.assertTrue(Instant.parse(modifiedTime).isAfter(before));
+    assertThat(Instant.parse(modifiedTime)).isGreaterThan(before);
   }
 
   @Test
@@ -134,12 +135,12 @@ public class UpdateCheckerTest {
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "1.0.2", testWebServer.getEndpoint(), "tool-name", ignored -> {});
-    Assert.assertTrue(message.isPresent());
-    Assert.assertEquals("2.0.0", message.get());
+
+    assertThat(message).hasValue("2.0.0");
     String modifiedTime =
         new String(
             Files.readAllBytes(configDir.resolve("lastUpdateCheck")), StandardCharsets.UTF_8);
-    Assert.assertTrue(Instant.parse(modifiedTime).isAfter(before));
+    assertThat(Instant.parse(modifiedTime)).isGreaterThan(before);
   }
 
   @Test
@@ -152,13 +153,14 @@ public class UpdateCheckerTest {
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "1.0.2", testWebServer.getEndpoint(), "tool-name", ignored -> {});
-    Assert.assertFalse(message.isPresent());
+
+    assertThat(message).isEmpty();
 
     // lastUpdateCheck should not have changed
     String lastUpdateTime =
         new String(
             Files.readAllBytes(configDir.resolve("lastUpdateCheck")), StandardCharsets.UTF_8);
-    Assert.assertEquals(Instant.parse(lastUpdateTime), modifiedTime.toInstant());
+    assertThat(modifiedTime.toInstant()).isEqualTo(Instant.parse(lastUpdateTime));
   }
 
   @Test
@@ -172,9 +174,9 @@ public class UpdateCheckerTest {
     String modifiedTime =
         new String(
             Files.readAllBytes(configDir.resolve("lastUpdateCheck")), StandardCharsets.UTF_8);
-    Assert.assertTrue(Instant.parse(modifiedTime).isAfter(before));
-    Assert.assertTrue(message.isPresent());
-    Assert.assertEquals("2.0.0", message.get());
+
+    assertThat(Instant.parse(modifiedTime)).isGreaterThan(before);
+    assertThat(message).hasValue("2.0.0");
   }
 
   @Test
@@ -190,11 +192,10 @@ public class UpdateCheckerTest {
               badServer.getEndpoint(),
               "tool-name",
               logEvent -> {
-                Assert.assertEquals(Level.DEBUG, logEvent.getLevel());
-                MatcherAssert.assertThat(
-                    logEvent.getMessage(), CoreMatchers.containsString("Update check failed; "));
+                assertThat(logEvent.getLevel()).isEqualTo(LogEvent.Level.DEBUG);
+                assertThat(logEvent.getMessage()).contains("Update check failed; ");
               });
-      Assert.assertFalse(message.isPresent());
+      assertThat(message).isEmpty();
     }
   }
 
@@ -202,8 +203,7 @@ public class UpdateCheckerTest {
   public void testFinishUpdateCheck_success() {
     Future<Optional<String>> updateCheckFuture = Futures.immediateFuture(Optional.of("Hello"));
     Optional<String> result = UpdateChecker.finishUpdateCheck(updateCheckFuture);
-    Assert.assertTrue(result.isPresent());
-    Assert.assertEquals("Hello", result.get());
+    assertThat(result).hasValue("Hello");
   }
 
   @Test
@@ -213,7 +213,7 @@ public class UpdateCheckerTest {
     Mockito.when(updateCheckFuture.isDone()).thenReturn(false);
 
     Optional<String> result = UpdateChecker.finishUpdateCheck(updateCheckFuture);
-    Assert.assertFalse(result.isPresent());
+    assertThat(result).isEmpty();
   }
 
   private void setupLastUpdateCheck() throws IOException {