+++ jib-cli/CHANGELOG.md
@@ -6,8 +6,13 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added automatic update check. Jib CLI will now display a message if a new version is available. See the [privacy page](https://github.com/GoogleContainerTools/jib/blob/master/docs/privacy.md) for more details. ([#3165](https://github.com/GoogleContainerTools/jib/pull/3165))
+- Added `--image-metadata-out` option to specify JSON output file that should contain image metadata (image ID, digest, and tags) after build is complete. ([#3187](https://github.com/GoogleContainerTools/jib/pull/3187))
 
 ### Changed
+
+### Fixed
+
+## 0.5.0
 
 ### Fixed